agent_service/requirements.txt
langchain==0.3.26
langchain-openai
langchain-mcp-adapters==0.1.8
langgraph-cli[inmem]==0.3.3
langfuse==2.60.3
pydantic==2.11.7
pydantic-settings==2.10.1
